+++ PracticalTask2.13/vehicle/Truck.java
